requirements.txt
PyPDF2
chromaDB
google.generativeai
dotenv
